# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plandy API",
        "description": "Plandy 백엔드 API 명세서",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ai/chat": {
            "post": {
                "tags": [
                    "AI 연동"
                ],
                "summary": "AI 채팅 메시지 전송",
                "operationId": "d0784401fb910eed3831ea246e3bbde5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chat_room_id",
                                    "message"
                                ],
                                "properties": {
                                    "chat_room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "오늘 할 일을 추천해줘"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AI 채팅 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "인증"
                ],
                "summary": "사용자 로그인",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kim@plandy.kr"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "로그인 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "인증 실패",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "사용자 정보",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "kim@plandy.kr"
                    },
                    "name": {
                        "type": "string",
                        "example": "김철수"
                    },
                    "timezone": {
                        "type": "string",
                        "example": "Asia/Seoul"
                    },
                    "preferences": {
                        "properties": {
                            "theme": {
                                "type": "string",
                                "example": "light"
                            },
                            "notifications": {
                                "type": "boolean",
                                "example": true
                            },
                            "work_hours": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "09:00",
                                    "18:00"
                                ]
                            },
                            "break_duration": {
                                "type": "integer",
                                "example": 15
                            },
                            "language": {
                                "type": "string",
                                "example": "ko"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}